+++ fort-to-nite/Helpers/WeaponsLevels.swift
@@ -119,7 +119,7 @@ class FormatLevels {
             cell.backgroundColor = GradientColor(UIGradientStyle.diagonal, frame: cell.frame, colors: [HexColor("4FCA00")!, HexColor("00BFFF")!])
             break
         case [1, 2, 3]:
-             cell.backgroundColor = GradientColor(UIGradientStyle.diagonal, frame: cell.frame, colors: [HexColor("4FCA00")!, HexColor("00BFFF")!, HexColor("B83DF2")!])
+             cell.backgroundColor = GradientColor(UIGradientStyle.diagonal, frame: cell.frame, colors: [HexColor("4FCA00")!, HexColor("4FCA00")!, HexColor("00BFFF")!, HexColor("B83DF2")!])
              break
         case [2, 3]:
             cell.backgroundColor = GradientColor(UIGradientStyle.diagonal, frame: cell.frame, colors: [HexColor("00BFFF")!, HexColor("B83DF2")!])